netwerk/cookie/test/mochitest/test_xmlDocument.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Document constructor</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<script type="application/javascript">

SpecialPowers.pushPrefEnv({set: [
  ["dom.storage_access.enabled", true],
]}).then(_ => {
  const doc = new Document();
  doc.requestStorageAccess().catch(__ => {}).then(___ => {
    ok(true, "No crash!");
    SimpleTest.finish();
  });
});

SimpleTest.waitForExplicitFinish();

</script>
</body>
</html>
